Moves XML parsing logic out of AWS.Util into AWS.XML
