+++ test/aws/util_test.exs
@@ -24,50 +24,4 @@ defmodule AWS.UtilTest do
     expected = "315f5bdb76d078c43b8ac0064e4a0164612b1fce77c869345bfc94c75894edd3"
     assert ^expected = Util.sha256_hexdigest("Hello, world!")
   end
-
-  @text "__text"
-
-  test "decode_xml/1 handles lists correctly by merging values in a list" do
-    expected = %{ "person" => %{ "name" => "foo",
-                                 "addresses" => %{"address" => ["1", "2"]}
-                               }
-                }
-    input = """
-    <person>
-      <name>foo</name>
-      <addresses>
-        <address>1</address>
-        <address>2</address>
-      </addresses>
-    </person>
-    """
-    assert ^expected = Util.decode_xml(input)
-  end
-
-  test "decode_xml/1 handles multiple text elments mixed with other elements correctly" do
-    expected = %{"person" => %{"name" => "foo", @text => "random"}}
-    input = """
-    <person>
-      <name>foo</name>
-      random
-    </person>
-    """
-    assert ^expected = Util.decode_xml(input)
-
-    expected = %{"person" => %{ "name" => "foo",
-                                "age" => "42",
-                                @text => "random\n  \n  text"
-                              }
-                }
-    input = """
-    <person>
-      <name>foo</name>
-      random
-      <age>42</age>
-      text
-    </person>
-    """
-    assert ^expected = Util.decode_xml(input)
-  end
-
 end
